i2c: Add an SMBus alert device.
This adds a device that conforms to the SMBus alert extension.  This
allows other I2C devices to register with this device so that when
they have data ready, it will interrupt the processor and allow the
processor to read from the alert device to see which i2c device has
data.

To use this, you must first create a named irq and then create
this device with:
  -device smbus-alert,name=<name>,irqname=<irqname>,address=..,bus=..
where the name is the value that other I2C devices will use to attach
to this device and the irqname is the name of the named irq that will
be raised when the data is ready.

Signed-off-by: Corey Minyard <cminyard@mvista.com>