tcg: Allow profiling stats to be put into share memory
For linux-user...

If -tcgkey <key> is set, put the TCG profiling structure into a system V
shared memory area for external access.

Also add the size of the table_op_count array to the structure because
it is machine-dependent.

Add a qemu-tcg program that is able to dump the TCG profiling data.

Signed-off-by: Corey Minyard <cminyard@mvista.com>