tcg: Rework profiling time calculations
The profile times for la_time and opt_time were being done in an unusual
manner that would result in bogus results if the measurements were taken
while the calculation was being done or the application terminated while
the calculation was being done.

Instead, grab a start time and end time and add the time between them
instead of subtracting the current time at the beginning and adding on
the current time at the end.

Signed-off-by: Corey Minyard <cminyard@mvista.com>