accel: Fix tb_lookup_cmp calculations
tb_lookup_cmp() needs to match tb_cmp(), but the tb_cflags were not
being compared in quite the same way.

Also, pre-calculate the second page address and put it in an array that
matches what is in the TranslationBlock structure so it's easier to
read and the second page doesn't need to be recalcuated on every hit.

Signed-off-by: Corey Minyard <cminyard@mvista.com>